break up big converter
